+++ server/tests/metadata.rs
@@ -7,6 +7,15 @@ mod entities_metadata_query {
     #[actix_rt::test]
     async fn returns_metadata_for_articles() {
         assert_metadata(json!({
+          "about": [
+            {
+              "type": "Concept",
+              "id": "http://w3id.org/kim/schulfaecher/s1017",
+              "inScheme": {
+                "id": "http://w3id.org/kim/schulfaecher/"
+              }
+            }
+          ],
           "@context": [
             "https://w3id.org/kim/amb/context.jsonld",
             {
@@ -172,6 +181,15 @@ mod entities_metadata_query {
     #[actix_rt::test]
     async fn returns_metadata_for_applets() {
         assert_metadata(json!({
+          "about": [
+            {
+              "type": "Concept",
+              "id": "http://w3id.org/kim/schulfaecher/s1017",
+              "inScheme": {
+                "id": "http://w3id.org/kim/schulfaecher/"
+              }
+            }
+          ],
           "@context": [
             "https://w3id.org/kim/amb/context.jsonld",
             {
@@ -249,6 +267,15 @@ mod entities_metadata_query {
     #[actix_rt::test]
     async fn returns_metadata_for_courses() {
         assert_metadata(json!({
+          "about": [
+            {
+              "type": "Concept",
+              "id": "http://w3id.org/kim/schulfaecher/s1017",
+              "inScheme": {
+                "id": "http://w3id.org/kim/schulfaecher/"
+              }
+            }
+          ],
           "@context": [
             "https://w3id.org/kim/amb/context.jsonld",
             {
@@ -345,6 +372,15 @@ mod entities_metadata_query {
     #[actix_rt::test]
     async fn returns_metadata_for_exercises() {
         assert_metadata(json!({
+          "about": [
+            {
+              "type": "Concept",
+              "id": "http://w3id.org/kim/schulfaecher/s1017",
+              "inScheme": {
+                "id": "http://w3id.org/kim/schulfaecher/"
+              }
+            }
+          ],
           "@context": [
             "https://w3id.org/kim/amb/context.jsonld",
             {
@@ -420,6 +456,15 @@ mod entities_metadata_query {
     #[actix_rt::test]
     async fn returns_metadata_for_exercise_groups() {
         assert_metadata(json!({
+          "about": [
+            {
+              "type": "Concept",
+              "id": "http://w3id.org/kim/schulfaecher/s1017",
+              "inScheme": {
+                "id": "http://w3id.org/kim/schulfaecher/"
+              }
+            }
+          ],
           "@context": [
             "https://w3id.org/kim/amb/context.jsonld",
             {
@@ -499,6 +544,15 @@ mod entities_metadata_query {
     #[actix_rt::test]
     async fn returns_metadata_for_videos() {
         assert_metadata(json!({
+          "about": [
+            {
+              "type": "Concept",
+              "id": "http://w3id.org/kim/schulfaecher/s1017",
+              "inScheme": {
+                "id": "http://w3id.org/kim/schulfaecher/"
+              }
+            }
+          ],
           "@context": [
             "https://w3id.org/kim/amb/context.jsonld",
             {